Refactor the index into separate files
